# api/http/openapi.yaml
openapi: 3.1.0
info:
  title: MatchLab Gateway API
  version: 2.3.0
  description: |
    Текущий публичный HTTP API MatchLab: учётные записи, сеансы, подтверждение почты и файлы.
    Это описание реализованных маршрутов Gateway. Для локальной проверки сначала
    зарегистрируйтесь, подтвердите email из письма, затем войдите и укажите полученный токен доступа в Bearer Auth.
    Для проверки почты сначала добавьте разрешённые домены HSE локальным seed-скриптом.
    Письмо с подтверждением появится в Mailpit по адресу http://localhost:8025.
    Токен из ссылки в письме передайте в метод подтверждения адреса.
servers:
  - url: http://localhost:8080
    description: Gateway в локальном Docker Compose
...
paths:
  /api/v1/media/files:
    post:
      tags: [Файлы]
      summary: Подготовить загрузку файла
      operationId: createMediaUpload
      description: |
        Требует сеанс и подтверждённую почту. Владелец определяется из сеанса.
        Возвращает pending и подписанный PUT. Передайте байты файла напрямую по upload_url,
        с выданными headers, затем вызовите complete. Не используйте multipart/form-data.
        В браузере Content-Length формируется автоматически. is_public по умолчанию false.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMediaUpload'
            example:
              original_name: avatar.png
              content_type: image/png
              size_bytes: 1024
              is_public: true
      responses:
        '201':
          description: Создана запись pending. Файл ещё не загружен.
          headers:
            Location:
              schema: {type: string}
              description: HTTP-путь к метаданным файла.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUpload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/MediaForbidden'
        '409':
          $ref: '#/components/responses/MediaConflict'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '504':
          $ref: '#/components/responses/MediaTimeout'
...
  /api/v1/media/files/multipart:
    post:
      tags: [Файлы]
      summary: Начать загрузку файла частями
      operationId: createMediaMultipart
      description: |
        Требует сеанс и подтверждённую почту. Возвращает pending-файл и параметры сессии.
        Размер части задаёт Media; последняя часть может быть меньше. Байты передаются
        напрямую в S3 через part-urls. Существующая загрузка одним PUT также доступна.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateMediaUpload'}
            example: {original_name: report.pdf, content_type: application/pdf, size_bytes: 8388611, is_public: false}
      responses:
        '201':
          description: Созданы pending-файл и сессия загрузки.
          headers:
            Location:
              schema: {type: string}
              description: Путь к состоянию multipart-сессии.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MediaMultipartState'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/MediaForbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
        '504': {$ref: '#/components/responses/MediaTimeout'}
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Токен access_token из ответа POST /api/v1/auth/login.
  schemas:
    MediaMultipartState:
      type: object
      required: [file, status, expected_size_bytes, part_size_bytes, part_count, expires_at_unix, uploaded_bytes, parts]
      properties:
        file: {$ref: '#/components/schemas/MediaFile'}
        status: {type: string, enum: [uploading, completing, completed, aborted]}
        expected_size_bytes: {type: integer, format: int64, minimum: 1}
        part_size_bytes: {type: integer, format: int64, minimum: 5242880, maximum: 5368709120}
        part_count: {type: integer, format: int32, minimum: 1, maximum: 10000}
        expires_at_unix: {type: integer, format: int64}
        uploaded_bytes: {type: integer, format: int64, minimum: 0}
        parts:
          type: array
          items:
            type: object
            required: [part_number, etag, size_bytes]
            properties:
              part_number: {type: integer, format: int32}
              etag: {type: string}
              size_bytes: {type: integer, format: int64}
...
    CreateMediaUpload:
      type: object
      additionalProperties: false
      required: [original_name, content_type, size_bytes]
      properties:
        original_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Имя без пути и управляющих символов.
        content_type:
          type: string
          description: Разрешённый MIME-тип без параметров; список задаётся настройками Media.
          examples: [application/pdf, image/jpeg, image/png]
        size_bytes:
          type: integer
          format: int64
          minimum: 1
          description: Размер в байтах; лимит из конфигурации Media (локально 100 MiB).
        is_public:
          type: boolean
          default: false
    MediaFile:
      type: object
      required: [id, owner_user_id, original_name, content_type, status, created_at_unix, is_public]
      properties:
        id: {type: string, format: uuid}
        owner_user_id: {type: string, format: uuid}
        original_name: {type: string}
        content_type: {type: string}
        size_bytes:
          type: integer
          format: int64
          description: Фактический размер; до подтверждения поле отсутствует.
        status:
          type: string
          enum: [pending, ready, deleting, deleted, failed]
        created_at_unix: {type: integer, format: int64}
        uploaded_at_unix:
          type: integer
          format: int64
          description: Время подтверждения в секундах Unix; до подтверждения отсутствует.
        is_public: {type: boolean}
...
    MediaUpload:
      type: object
      required: [file, upload_url, method, headers, expires_at_unix]
      properties:
        file:
          $ref: '#/components/schemas/MediaFile'
        upload_url: {type: string, format: uri}
        method: {type: string, enum: [PUT]}
        headers:
          type: object
          additionalProperties: {type: string}
          description: Заголовки подписанного PUT. Content-Length в браузере задаёт сам браузер.
        expires_at_unix: {type: integer, format: int64}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
...
    MediaConflict:
      description: Неподходящее состояние, объект ещё не загружен, неверные метаданные объекта или дубликат.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MediaForbidden:
      description: Подтвердите почту или доступ запрещён.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MediaTimeout:
      description: Превышено время ожидания ответа Media.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Внутренняя ошибка без подробностей реализации.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Некорректный JSON или значения запроса.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: некорректное тело запроса
    Unauthorized:
      description: Токен доступа отсутствует или недействителен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: недействительный сеанс
...
    Unavailable:
      description: Зависимый сервис или SMTP-сервер недоступен.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: сервис недоступен